rename src/v3-spec to src/build-v3-spec
